Improved API calls, observers, added exceptions, added info line

+++ Frontend/src/main/java/at/fhtw/swen2_tourplanner/frontend/controller/TourLogDataController.java
@@ -24,7 +24,7 @@ public class TourLogDataController extends BaseController<TourLogData> {
     TableColumn<TourLog, Double> ratingCol;
 
     @FXML
-    TableColumn<TourLog, Long> distanceCol;
+    TableColumn<TourLog, Double> distanceCol;
     @FXML
     private TableView<TourLog> logTableView;
